requirements.txt
appdirs==1.4.4
beautifulsoup4==4.9.1
bokeh==2.2.1
branca==0.3.1
bs4==0.0.1
certifi==2018.11.29
chardet==3.0.4
click==7.1.2
cssselect==1.1.0
fake-useragent==0.1.11
Flask==1.1.2
Flask-WTF==0.14.3
folium==0.7.0
get-all-tickers==1.7
gunicorn==20.0.4
idna==2.8
itsdangerous==1.1.0
Jinja2==2.11.2
lxml==4.5.2
MarkupSafe==1.1.0
mysql-connector-python==8.0.21
numpy==1.16.1
packaging==20.4
pandas==1.1.2
pandas-datareader==0.9.0
parse==1.18.0
Pillow==7.2.0
protobuf==3.13.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
pyee==7.0.4
pyparsing==2.4.7
pyppeteer==0.2.2
pyquery==1.4.1
python-dateutil==2.8.0
pytz==2018.9
PyYAML==5.3.1
requests==2.21.0
six==1.12.0
soupsieve==2.0.1
tornado==5.1
tqdm==4.49.0
typing-extensions==3.7.4.3
urllib3==1.24.1
w3lib==1.22.0
websockets==8.1
Werkzeug==1.0.1
WTForms==2.3.3
